requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
orpheus-speech
vllm==0.7.3
librosa>=0.10.0
soundfile>=0.12.0
transformers>=4.30.0
datasets>=2.10.0
wandb>=0.15.0
accelerate>=0.20.0
matplotlib>=3.5.0
scipy>=1.9.0
numpy>=1.21.0
PyYAML>=6.0
ipywidgets
jupyter
